src/main/java/com/project/account/service/CustomerService.java
package com.project.account.service;

import com.project.account.dto.CustomerDto;
import com.project.account.dto.CustomerDtoConverter;
import com.project.account.exception.CustomerNotFoundException;
import com.project.account.model.Customer;
import com.project.account.repository.CustomerRepository;
import org.springframework.stereotype.Service;

@Service
public class CustomerService {

    private final CustomerRepository customerRepository;
    private final CustomerDtoConverter converter;

    public CustomerService(CustomerRepository customerRepository, CustomerDtoConverter converter) {
        this.customerRepository = customerRepository;
        this.converter = converter;
    }

    protected Customer findCustomerById(String id) {
        return customerRepository.findById(id).orElseThrow(
                () -> new CustomerNotFoundException("Customer could not find by id:" +id));

    }

    public CustomerDto getCustomerById(String customerId) {
        return converter.convertToCustomerDto(findCustomerById(customerId));
    }
}
